requirements.txt
langchain==0.2.0
langgraph==0.0.40
langchain-community==0.2.0
langchain-google-genai==1.0.2
google-generativeai==0.5.4
gradio==4.26.0
pandas==2.2.2
python-dotenv==1.0.1
